autogen_demo/requirements.txt
autogen==0.4.2b1
autogen==0.4.2b1
autogen-agentchat==0.6.4
autogen-core==0.6.4
autogen-ext==0.6.4
gradio>=4.14.0
requests==2.31.0
httpx==0.26.0
pandas==2.2.0
pydantic==2.5.0
websockets==12.0
beautifulsoup4==4.12.0
openai==1.14.0 
tenacity
python-dotenv 